add checkout view
