+++ src/helpers/routes.ts
@@ -22,6 +22,7 @@ export const FAQ = '/faq'
 export const ITEMS = '/items'
 export const items = (policyId: string) => `/policies/${policyId}/items`
 export const itemsNew = (policyId: string) => `/policies/${policyId}/items/new`
+export const itemsCheckout = (policyId: string, ItemId: string) => `/policies/${policyId}/items/${ItemId}/checkout`
 export const itemDetails = (policyId: string, itemId: string) => `/policies/${policyId}/items/${itemId}`
 export const itemEdit = (policyId: string, itemId: string) => `/policies/${policyId}/items/${itemId}/edit`
 export const itemNewClaim = (policyId: string, itemId: string) => `/policies/${policyId}/items/${itemId}/new-claim`
