+++ src/pages/policies/[policyId]/items/new.svelte
@@ -1,19 +1,17 @@
 <script lang="ts">
-import Checkout from 'Checkout.svelte'
 import { Breadcrumb, ItemForm, NoHouseholdIdModal } from 'components'
 import { loadDependents } from 'data/dependents'
-import { addItem, deleteItem, loadItems, PolicyItem, submitItem } from 'data/items'
+import { addItem, loadItems, PolicyItem } from 'data/items'
 import { PolicyType, selectedPolicy, updatePolicy } from 'data/policies'
 import { loadMembersOfPolicy } from 'data/policy-members'
 import { formatPageTitle } from 'helpers/pageTitle'
-import { HOME, items as itemsRoute, itemDetails, itemsNew } from 'helpers/routes'
+import { HOME, items as itemsRoute, itemsCheckout, itemsNew } from 'helpers/routes'
 import { goto, metatags } from '@roxi/routify'
 import { Page, setNotice } from '@silintl/ui-components'
 import { onMount } from 'svelte'
 
 export let policyId: string
 
-let isCheckingOut = false
 let item: PolicyItem
 let open = false
 
@@ -35,28 +33,13 @@ $: breadcrumbLinks = [
 
 const onApply = async (event: CustomEvent) => {
   item = await addItem(policyId, event.detail)
-  isCheckingOut = true
+  $goto(itemsCheckout(policyId, item.id))
 }
 
 const onSaveForLater = async (event: CustomEvent) => {
   await addItem(policyId, event.detail)
 
   $goto(HOME)
-}
-
-const onAgreeAndPay = async (event: CustomEvent<string>) => {
-  const itemId = event.detail
-  await submitItem(itemId)
-  $goto(itemDetails(policyId, itemId))
-}
-
-const onDelete = async (event: CustomEvent<string>) => {
-  await deleteItem(policyId, event.detail)
-  $goto(itemsRoute(policyId))
-}
-
-const onEdit = () => {
-  isCheckingOut = false
 }
 
 const onClosed = async (event: CustomEvent<any>) => {
@@ -72,11 +55,7 @@ const onClosed = async (event: CustomEvent<any>) => {
 </script>
 
 <Page>
-  {#if !isCheckingOut}
-    <Breadcrumb links={breadcrumbLinks} />
-    <ItemForm {item} {policyId} on:submit={onApply} on:save-for-later={onSaveForLater} />
-    <NoHouseholdIdModal {open} on:closed={onClosed} />
-  {:else}
-    <Checkout {item} {policyId} on:agreeAndPay={onAgreeAndPay} on:delete={onDelete} on:edit={onEdit} />
-  {/if}
+  <Breadcrumb links={breadcrumbLinks} />
+  <ItemForm {item} {policyId} on:submit={onApply} on:save-for-later={onSaveForLater} />
+  <NoHouseholdIdModal {open} on:closed={onClosed} />
 </Page>
